+++ src/Gaffer/Animation.cpp
@@ -397,8 +397,10 @@ private:
 
 	double solveForTime( const double tl, const double th, const double time ) const
 	{
-		if( time <= 0.0 ) return 0.0;
-		if( time >= 1.0 ) return 1.0;
+		// NOTE : keeping tl and th in the range [0,1] ensures f is monotonic increasing over interval [0,1].
+
+		assert( 0.0 <= tl && tl <= 1.0 );
+		assert( 0.0 <= th && th <= 1.0 );
 
 		// compute coeffs
 
@@ -409,26 +411,21 @@ private:
 		const double bt2 = bt + bt;
 		const double at3 = at + at + at;
 
-		// check that f is monotonic and therefore has one (possibly repeated) real root.
+		// check that f is monotonic increasing over interval [0,1] and therefore has one (possibly repeated) real root.
 		//
-		// NOTE : f is monotonic over the interval [0,1] when the solutions of f' either,
-		//        both lie outside the interval (0,1) or lie in the interval (0,1) and are equal
-		//        in which case the discriminant of f' is zero.
-		// NOTE : keeping tl and th in the range [0,1] ensures f is monotonic over interval [0,1].
+		// NOTE : As f(0) = 0 and f(1) = 1, f is monotonic increasing iff f'(0) >= 0 and f'(1) >= 0
+		//
+		//        f'(0) =                 c(t)
+		//        f'(1) = 3a(t) + 2b(t) + c(t)
+		//
+		//        when th == 1 floating point imprecision gives f'(1) as slighty less than 0.
 
-		const double discriminant = bt2 * bt2 - 4.0 * at3 * ct;
+		assert( ( ct >= 0.0 ) && ( at3 + bt2 + ct >= ( ( th == 1.0 ) ? -1e-15 : 0.0 ) ) );
 
-		if( discriminant > 1e-13 )
-		{
-			const double q = - 0.5 * ( bt2 + std::copysign( std::sqrt( discriminant ), bt2 ) );
-			const double s1 = q / at3;
-			const double s2 = ct / q;
+		// simple cases
 
-			if( ( 0.0 < s1 && s1 < 1.0 ) || ( 0.0 < s2 && s2 < 1.0 ) )
-			{
-				throw IECore::Exception( "Animation : Bezier interpolation mode : curve segment has multiple values for given time." );
-			}
-		}
+		if( time <= 0.0 ) return 0.0;
+		if( time >= 1.0 ) return 1.0;
 
 		// root bracketed in interval [0,1]