+++ Changes.md
@@ -4,7 +4,8 @@
 Fixes
 -----
 
-- Animation Gadget : Fixed bug that prevented signals from being disconnected when key deleted.
+- Animation : Prevent negative zero slope.
+- AnimationGadget : Fixed bug that prevented signals from being disconnected when key deleted.
 
 0.61.14.1 (relative to 0.61.14.0)
 =========