+++ Changes.md
@@ -11,6 +11,8 @@ Fixes
 -----
 
 - ImageWriter : Fix bug with color space unpremulting when alpha channel not the last channel.
+- Viewer : Fix images not displaying when display window is offscreen.
+- UV Inspector : Fix texture display ( broken in 1.0.0.0 ).
 
 1.0.1.0 (relative to 1.0.0.0)
 =======